+++ src/olcs/FeatureConverter.js
@@ -10,7 +10,16 @@ import {boundingExtent, getCenter} from 'ol/extent.js';
 import olGeomSimpleGeometry from 'ol/geom/SimpleGeometry.js';
 import {convertColorToCesium, olGeometryCloneTo4326, ol4326CoordinateToCesiumCartesian, ol4326CoordinateArrayToCsCartesians} from './core.ts';
 import olcsCoreVectorLayerCounterpart from './core/VectorLayerCounterpart.ts';
-import {obj, getUid, isGroundPolylinePrimitiveSupported, waitReady} from './util.js';
+import {getUid, isGroundPolylinePrimitiveSupported, waitReady} from './util.ts';
+
+/**
+ * Cast to object.
+ * @param {Object} param
+ * @return {Object}
+ */
+function obj(param) {
+  return param;
+}
 
 
 /**